refactor: move localization helpers into shared scope
