+++ EssentialApp/EssentialAppTests/Helpers/SharedTestHelpers.swift
@@ -17,3 +17,12 @@ func anyData() -> Data { Data("any data".utf8) }
 func uniqueFeed() -> [FeedImage] {
     [FeedImage(id: .init(), description: "any", location: "any", url: anyURL())]
 }
+
+private class DummyView: ResourceView {
+    func display(_ viewModel: Any) { }
+}
+
+var loadError: String { LoadResourcePresenter<Any, DummyView>.loadError }
+var feedTitle: String { FeedPresenter.title }
+
+var commentsTitle: String { ImageCommentsPresenter.title }
